camera_webapp/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
grpcio==1.63.0
grpcio-tools==1.63.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
protobuf==5.26.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==69.5.1
SQLAlchemy==2.0.25
typing_extensions==4.11.0
Werkzeug==3.0.3
